feat: fix app metadata
