+++ src/client/components/common/seo.tsx
@@ -1,15 +1,53 @@
 import Head from 'next/head';
 import { FunctionComponent } from 'react';
-import { PRODUCTION_URL } from './content';
+import { PRODUCTION_URL, UseCase } from './content';
+import { Metadata } from 'next';
 
-type HeadSEOProps = {
+type SeoProps = {
   title: string;
   description: string;
   image?: string;
   path?: string;
 };
 
-export const SEO: FunctionComponent<HeadSEOProps> = ({ title, description, image, path }) => {
+/**
+ * Generates Metadata object for Next `app` directory
+ */
+export const generateMetadata = ({ title, description, image, path }: SeoProps): Metadata => {
+  const metaImage = image ?? `${PRODUCTION_URL}/fingerprintDefaultMetaImage.png`;
+  const metaUrl = `${PRODUCTION_URL}${path ?? ''}`;
+  return {
+    title,
+    description,
+    openGraph: {
+      type: 'website',
+      title,
+      description,
+      siteName: 'Fingerprint Use Cases',
+      images: [metaImage],
+      url: metaUrl,
+    },
+    twitter: {
+      title,
+      description,
+      card: 'summary_large_image',
+      images: [metaImage],
+      site: metaUrl,
+    },
+  };
+};
+
+export const generateUseCaseMetadata = (useCase: UseCase): Metadata =>
+  generateMetadata({
+    title: useCase.titleMeta,
+    description: useCase.descriptionMeta,
+    path: useCase.url,
+  });
+
+/**
+ * Generates next/Head tags for Next `pages` directory
+ */
+export const SEO: FunctionComponent<SeoProps> = ({ title, description, image, path }) => {
   const metaImage = image ?? `${PRODUCTION_URL}/fingerprintDefaultMetaImage.png`;
   const metaUrl = `${PRODUCTION_URL}${path ?? ''}`;
   return (
